機能改修はせずに、コンポーネントわけ、リファクタリング、コメントの追加を実施。
